Create 03Pass-by-Value
The compiler never actually sends the variables into the function. Instead, a copy of each argument is made and used in the scope of the function. This concept is known as passing-by-value. Only the values of the arguments are passed into the functions, not the arguments themselves.
When the function ends, all the argument copies and the variables created inside the function are destroyed forever. Basically, changing the value of an argument inside a function won’t change it outside the function.